WIP e2e and service tests
