+++ apps/account-api/src/services/siwfV2.mock.spec.ts
@@ -7,7 +7,13 @@ import { URL } from 'node:url';
 // validSiwfAddDelegationResponsePayload
 // validSiwfLoginResponsePayload
 // validSiwfNewUserResponse
+// validSiwfNewUserResponseWithRecovery
 // Anything else? 404
+
+// These are payloads for testing signing in/up with Frequency Access.
+// The authorization codes are mocked to be the same
+// name as the payloads below.  Frequency Access would be generating random
+// authorization codes in production.
 
 export const validSiwfAddDelegationResponsePayload: SiwfResponse = {
   userPublicKey: {
@@ -304,10 +310,155 @@ export const validSiwfNewUserResponse: SiwfResponse = {
   ],
 };
 
+export const validSiwfNewUserResponseWithRecovery: SiwfResponse = {
+  userPublicKey: {
+    encodedValue: 'f6akufkq9Lex6rT8RCEDRuoZQRgo5pWiRzeo81nmKNGWGNJdJ',
+    encoding: 'base58',
+    format: 'ss58',
+    type: 'Sr25519',
+  },
+  payloads: [
+    {
+      signature: {
+        algo: 'SR25519',
+        encoding: 'base16',
+        encodedValue:
+          '0xcc469ce5304768c16164e42442dfe40769a742786bd360cabb8fdaf15a25765c9c18d71b8c88972f42ab6f938e703f4cc4035656a7c9dd193aca2626a3d7f887',
+      },
+      endpoint: {
+        pallet: 'msa',
+        extrinsic: 'createSponsoredAccountWithDelegation',
+      },
+      type: 'addProvider',
+      payload: {
+        authorizedMsaId: 1,
+        schemaIds: [5, 7, 8, 9, 10],
+        expiration: 24,
+      },
+    },
+    {
+      signature: {
+        algo: 'SR25519',
+        encoding: 'base16',
+        encodedValue:
+          '0x90583e8c2bdb46afd62caed94eb2f6c99f66308a91591da1447d2ef781002622ef12b960fc86de88121e243ad3cf5e15a615d48e7c2e20c860a7adf6199cfb88',
+      },
+      endpoint: {
+        pallet: 'statefulStorage',
+        extrinsic: 'applyItemActionsWithSignatureV2',
+      },
+      type: 'itemActions',
+      payload: {
+        schemaId: 7,
+        targetHash: 0,
+        expiration: 20,
+        actions: [
+          {
+            type: 'addItem',
+            payloadHex: '0x40eea1e39d2f154584c4b1ca8f228bb49ae5a14786ed63c90025e755f16bd58d37',
+          },
+        ],
+      },
+    },
+    {
+      signature: {
+        algo: 'SR25519',
+        encoding: 'base16',
+        encodedValue:
+          '0xa291b22f6d3e21258200f127ea23b741248a584208d96783fb3599a733a87c2e67fe427491a6aa984e3982e5ccafcbb7f170a17a5db6082230a1c8e4ac2a3f8c',
+      },
+      endpoint: {
+        pallet: 'handles',
+        extrinsic: 'claimHandle',
+      },
+      type: 'claimHandle',
+      payload: {
+        baseHandle: 'ExampleHandle',
+        expiration: 24,
+      },
+    },
+    {
+      signature: {
+        algo: 'SR25519',
+        encoding: 'base16',
+        encodedValue:
+          '0x04084ae8ced4e9e27cbc27bf25a3c34cb6faeec5fbdcb331021d98dc4a211f17256c08aa96defcb597f3867b08b3a901d61c8ce36e0f4ccda0bf79458ad39580',
+      },
+      endpoint: {
+        pallet: 'msa',
+        extrinsic: 'addRecoveryCommitment',
+      },
+      type: 'recoveryCommitment',
+      payload: {
+        recoveryCommitmentHex: '0xeea1e39d2f154584c4b1ca8f228bb49ae5a14786ed63c90025e755f16bd58d37',
+        expiration: 20,
+      },
+    },
+  ],
+  credentials: [
+    {
+      '@context': ['https://www.w3.org/ns/credentials/v2', 'https://www.w3.org/ns/credentials/undefined-terms/v2'],
+      type: ['VerifiedEmailAddressCredential', 'VerifiableCredential'],
+      issuer: 'did:web:frequencyaccess.com',
+      validFrom: '2024-08-21T21:28:08.289+0000',
+      credentialSchema: {
+        type: 'JsonSchema',
+        id: 'https://schemas.frequencyaccess.com/VerifiedEmailAddressCredential/bciqe4qoczhftici4dzfvfbel7fo4h4sr5grco3oovwyk6y4ynf44tsi.json',
+      },
+      credentialSubject: {
+        id: 'did:key:z6QNucQV4AF1XMQV4kngbmnBHwYa6mVswPEGrkFrUayhttT1',
+        emailAddress: 'john.doe@example.com',
+        lastVerified: '2024-08-21T21:27:59.309+0000',
+      },
+      proof: {
+        type: 'DataIntegrityProof',
+        verificationMethod: 'did:web:frequencyaccess.com#z6MkofWExWkUvTZeXb9TmLta5mBT6Qtj58es5Fqg1L5BCWQD',
+        cryptosuite: 'eddsa-rdfc-2022',
+        proofPurpose: 'assertionMethod',
+        proofValue: 'z4jArnPwuwYxLnbBirLanpkcyBpmQwmyn5f3PdTYnxhpy48qpgvHHav6warjizjvtLMg6j3FK3BqbR2nuyT2UTSWC',
+      },
+    },
+    {
+      '@context': ['https://www.w3.org/ns/credentials/v2', 'https://www.w3.org/ns/credentials/undefined-terms/v2'],
+      type: ['VerifiedGraphKeyCredential', 'VerifiableCredential'],
+      issuer: 'did:key:z6QNucQV4AF1XMQV4kngbmnBHwYa6mVswPEGrkFrUayhttT1',
+      validFrom: '2024-08-21T21:28:08.289+0000',
+      credentialSchema: {
+        type: 'JsonSchema',
+        id: 'https://schemas.frequencyaccess.com/VerifiedGraphKeyCredential/bciqmdvmxd54zve5kifycgsdtoahs5ecf4hal2ts3eexkgocyc5oca2y.json',
+      },
+      credentialSubject: {
+        id: 'did:key:z6QNucQV4AF1XMQV4kngbmnBHwYa6mVswPEGrkFrUayhttT1',
+        encodedPublicKeyValue: '0xb5032900293f1c9e5822fd9c120b253cb4a4dfe94c214e688e01f32db9eedf17',
+        encodedPrivateKeyValue: '0xd0910c853563723253c4ed105c08614fc8aaaf1b0871375520d72251496e8d87',
+        encoding: 'base16',
+        format: 'bare',
+        type: 'X25519',
+        keyType: 'dsnp.public-key-key-agreement',
+      },
+    },
+    {
+      '@context': ['https://www.w3.org/ns/credentials/v2', 'https://www.w3.org/ns/credentials/undefined-terms/v2'],
+      type: ['VerifiedRecoverySecretCredential', 'VerifiableCredential'],
+      issuer: 'did:key:z6QNucQV4AF1XMQV4kngbmnBHwYa6mVswPEGrkFrUayhttT1',
+      validFrom: '2024-08-21T21:28:08.289+0000',
+      credentialSchema: {
+        type: 'JsonSchema',
+        id: 'https://schemas.frequencyaccess.com/VerifiedRecoverySecretCredential/bciqpg6qm4rnu2j4v6ghxqqgwkggokwvxs3t2bexbd3obkypkiryylxq.json',
+      },
+      credentialSubject: {
+        id: 'did:key:z6QNucQV4AF1XMQV4kngbmnBHwYa6mVswPEGrkFrUayhttT1',
+        recoverySecret: '69EC-2382-E1E6-76F3-341F-3414-9DD5-CFA5-6932-E418-9385-0358-31DF-AFEA-9828-D3B7',
+      },
+    },
+  ],
+};
+
 const responseMap = new Map([
   ['validSiwfAddDelegationResponsePayload', validSiwfAddDelegationResponsePayload],
   ['validSiwfLoginResponsePayload', validSiwfLoginResponsePayload],
   ['validSiwfNewUserResponse', validSiwfNewUserResponse],
+  ['validSiwfNewUserResponseWithRecovery', validSiwfNewUserResponseWithRecovery],
   ['validEthereumSiwfLoginResponsePayload', validEthereumSiwfLoginResponsePayload],
 ]);
