+++ libs/types/src/dtos/account/transaction.request.dto.ts
@@ -27,7 +27,6 @@ export type TransactionData<
     | PublishRetireMsaRequestDto
     | PublicKeyAgreementRequestDto
     | PublishRevokeDelegationRequestDto,
-  // TODO: | PublishAddRecoveryCommitmentDto
 > = RequestType & {
   providerId: string;
   referenceId: string;
