+++ libs/utils/src/decorators/is-account-id-address.decorator.ts
@@ -2,7 +2,7 @@ import { registerDecorator, ValidationArguments, ValidationOptions } from 'class
 import { encodeAddress } from '@polkadot/keyring';
 import { hexToU8a } from '@polkadot/util';
 
-// TODO: correct validations for Ethereum Addresses
+//  Needs cleanup.  See: https://github.com/ProjectLibertyLabs/gateway/issues/883
 export function IsAccountIdOrAddress(validationOptions?: ValidationOptions) {
   // eslint-disable-next-line func-names
   return function (object: object, propertyName: string) {
