+++ package.json
@@ -119,8 +119,8 @@
     "@dsnp/activity-content": "^1.1.0",
     "@dsnp/parquetjs": "^1.8.6",
     "@dsnp/schemas": "^1.3.2",
-    "@frequency-chain/api-augment": "^1.17.0",
-    "@frequency-chain/ethereum-utils": "^1.17.0",
+    "@frequency-chain/api-augment": "^1.17.1",
+    "@frequency-chain/ethereum-utils": "^1.17.1",
     "@hpke/core": "^1.7.2",
     "@hpke/dhkem-x25519": "^1.6.2",
     "@multiformats/blake2": "^2.0.2",
@@ -143,7 +143,7 @@
     "@polkadot/util-crypto": "^13.5.3",
     "@projectlibertylabs/graph-sdk": "^2.0.1",
     "@projectlibertylabs/siwfv1": "npm:@projectlibertylabs/siwf@^1.1.7",
-    "@projectlibertylabs/siwfv2": "npm:@projectlibertylabs/siwf@^2.3.1",
+    "@projectlibertylabs/siwfv2": "npm:@projectlibertylabs/siwf@^2.3.2",
     "@songkeys/nestjs-redis": "^10.0.0",
     "@types/express": "^5.0.0",
     "@types/multer": "^1.4.12",
